+++ CHANGELOG.md
@@ -7,10 +7,11 @@ All Notable changes to `Csv` will be documented in this file
 ### Added
 
 - `Statement::select`
+- `TabularDataReader::getRecordsAsObject`
 
 ### Deprecated
 
-- None
+- `TabularDataReader::getObjects` use `TabularDataReader::getRecordsAsObject` instead
 
 ### Fixed
 
@@ -18,6 +19,7 @@ All Notable changes to `Csv` will be documented in this file
 - `Statement` should not throw when `LimitIterator` is used in combinaison with `ArrayIterator`.
 - `Statement` internal codebase improvement.
 - Using the `$header` argument on `Statement::process` is no longer deprecated. `E_USER_DEPRECATED` is no longer triggered.
+- `ResultSet::getRecords` should not return `null` see [#514](https://github.com/thephpleague/csv/issues/514)
 
 ### Removed